+++ headless_doom/r_draw.c
@@ -472,7 +472,7 @@ void R_InitTranslationTables (void)
     int		i;
 	
     translationtables = Z_Malloc (256*3+255, PU_STATIC, 0);
-    translationtables = (byte *)(( (int)translationtables + 255 )& ~255);
+    translationtables = (byte *)(( (intptr_t)translationtables + 255 )& ~255); // JWh - 64-bit compat
     
     // translate just the 16 green colors
     for (i=0 ; i<256 ; i++)
